Move the error message to the PaymentSheetActivity and PaymentSheetViewModel
