+++ stripe/src/main/java/com/stripe/android/paymentsheet/PaymentOptionsViewModel.kt
@@ -154,4 +154,8 @@ internal class PaymentOptionsViewModel(
             ) as T
         }
     }
+
+    override fun onBackPressed() {
+        // No action required
+    }
 }
