+++ stripe/src/main/java/com/stripe/android/paymentsheet/ui/BaseSheetActivity.kt
@@ -11,7 +11,6 @@ import androidx.annotation.DrawableRes
 import androidx.annotation.StringRes
 import androidx.appcompat.app.AppCompatActivity
 import androidx.core.content.ContextCompat
-import androidx.core.view.isVisible
 import com.google.android.material.appbar.AppBarLayout
 import com.google.android.material.appbar.MaterialToolbar
 import com.stripe.android.R
@@ -70,11 +69,6 @@ internal abstract class BaseSheetActivity<ResultType> : AppCompatActivity() {
             if (shouldFinish) {
                 finish()
             }
-        }
-
-        viewModel.userErrorMessage.observe(this) { userMessage ->
-            messageView.isVisible = userMessage != null
-            messageView.text = userMessage?.message
         }
 
         viewModel.processing.observe(this) { isProcessing ->
